Move petab v1 simulation code to amici.sim...
Related to #3041.
