+++ python/sdist/amici/importers/petab/v1/__init__.py
@@ -5,31 +5,7 @@
 PREEQ_INDICATOR_ID = "preequilibration_indicator"
 
 from .petab_import import import_petab_problem
-from .simulations import (
-    EDATAS,
-    FIM,
-    LLH,
-    RDATAS,
-    RES,
-    S2LLH,
-    SLLH,
-    SRES,
-    rdatas_to_measurement_df,
-    rdatas_to_simulation_df,
-    simulate_petab,
-)
 
 __all__ = [
     "import_petab_problem",
-    "simulate_petab",
-    "rdatas_to_simulation_df",
-    "rdatas_to_measurement_df",
-    "LLH",
-    "SLLH",
-    "FIM",
-    "S2LLH",
-    "RES",
-    "SRES",
-    "RDATAS",
-    "EDATAS",
 ]
