+++ python/sdist/amici/sim/sundials/petab/v1/_conditions.py
@@ -19,20 +19,19 @@ from petab.v1.C import (
     TIME,
 )
 
-from amici.sim.sundials import (
-    AmiciModel,
-    ExpData,
-    parameter_scaling_from_int_vector,
-)
-
-from .parameter_mapping import (
+from amici.importers.petab.v1.parameter_mapping import (
     ParameterMapping,
     ParameterMappingForCondition,
     petab_to_amici_scale,
     scale_parameters_dict,
     unscale_parameters_dict,
 )
-from .util import get_states_in_condition_table
+from amici.importers.petab.v1.util import get_states_in_condition_table
+from amici.sim.sundials import (
+    AmiciModel,
+    ExpData,
+    parameter_scaling_from_int_vector,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -283,7 +282,9 @@ def create_parameterized_edatas(
 
     # Get parameter mapping
     if parameter_mapping is None:
-        from .parameter_mapping import create_parameter_mapping
+        from amici.importers.petab.v1.parameter_mapping import (
+            create_parameter_mapping,
+        )
 
         parameter_mapping = create_parameter_mapping(
             petab_problem=petab_problem,
