+++ python/sdist/amici/sim/sundials/petab/v1/_simulations.py
@@ -17,6 +17,10 @@ import petab.v1 as petab
 from petab.v1.C import *  # noqa: F403
 
 import amici
+from amici.importers.petab.v1.parameter_mapping import (
+    ParameterMapping,
+    create_parameter_mapping,
+)
 from amici.logging import get_logger, log_execution_time
 from amici.sim.sundials import (
     AMICI_SUCCESS,
@@ -27,14 +31,9 @@ from amici.sim.sundials import (
     SensitivityOrder,
     run_simulations,
 )
-
-from .conditions import (
+from amici.sim.sundials.petab.v1._conditions import (
     create_edatas,
     fill_in_parameters,
-)
-from .parameter_mapping import (
-    ParameterMapping,
-    create_parameter_mapping,
 )
 
 logger = get_logger(__name__)
